REGRESSION (283084@main): Document::visibilityStateChanged does not hold reference to callback clients rdar://138799302 https://bugs.webkit.org/show_bug.cgi?id=282360
Reviewed by NOBODY (OOPS!).

Document::visibilityStateChanged() invokes visibility state callback clients, but does not
hold a reference to them before invoking. The client could then accidentally free itself
and cause an UAF. One possible route that leads to an UAF is through ViewTransition,
which is used in the test case:

* The ViewTransition C++ objects are allocated by document.startViewTransition().
  After the call, each object has a ref count of at least 2 (one in the JS wrapper
  that wraps the C++ object, one in Document::m_activeViewTransition)
* The GC is invoked, which releases the JS wrappers and decreases the ref count to 1
* The document visibility state is changed. This invokes ViewTransition::visibilityStateChanged
  on each object, which calls ::skipViewTransition, which calls ::clearViewTransition.
  ::clearViewTransition sets Document::m_activeViewTransition to null, so the object ref
  count is 0 and it's deallocated. ::clearViewTransition then continues to modify the
  (already deallocated) object, leading to an UAF.

Fix this by holding a reference to the callback clients before invoking it. This involves
making VisibilityChangeClient ref counted. This has the side effect of requiring
WakeLockManager (which inherits VisibilityChangeClient) to be made ref counted. Then
Document::visibilityStateChanged() would hold the client in a Ref before invoking it.

It's also observed that the JS wrapper should not be deallocated by the GC before the
view transition has completed. This commit fixes this by implementing
ViewTransition::virtualHasPendingActivity(), which the GC consults to determine whether
to deallocate the wrapper or not.

* LayoutTests/fast/dom/view-transition-lifetime-crash-expected.txt: Added.
* LayoutTests/fast/dom/view-transition-lifetime-crash.html: Added.
* Source/WebCore/Modules/screen-wake-lock/WakeLockManager.h:
* Source/WebCore/dom/Document.cpp:
(WebCore::Document::visibilityStateChanged):
(WebCore::Document::wakeLockManager):
* Source/WebCore/dom/ViewTransition.cpp:
(WebCore::ViewTransition::virtualHasPendingActivity const):
* Source/WebCore/dom/ViewTransition.h:
* Source/WebCore/dom/VisibilityChangeClient.h: