Suppress "refcounted class without virtual destructor" warning in derived classes of StyleRuleBase rdar://141314048 https://bugs.webkit.org/show_bug.cgi?id=284489
Reviewed by NOBODY (OOPS!).

StyleRuleBase defines a custom "destroying delete" operator, which calls the
appropriate derived type destructor before deallocating the object. This
avoid the need for virtual destructor and improves performance/memory usage,
as there's no vtable to store and consult. However the safer C++ checker doesn't
know about this and throw the "refcounted class without virtual destructor"
warning, so suppress it. This has to be done for each derived classes of
StyleRuleBase.

* Source/WebCore/css/CSSCounterStyleRule.h:
* Source/WebCore/css/CSSKeyframeRule.h:
* Source/WebCore/css/CSSKeyframesRule.h:
* Source/WebCore/css/CSSPositionTryRule.h:
* Source/WebCore/css/CSSViewTransitionRule.h:
* Source/WebCore/css/StyleRule.h:
* Source/WebCore/css/StyleRuleImport.h: