Boxes in the top layer always use the initial containing block as their static-position rectangle https://bugs.webkit.org/show_bug.cgi?id=295686
Reviewed by NOBODY (OOPS!).

When an absolutely-positioned box uses "inset: auto", its insets are
resolved such that the box aligns with its static position. This works
for boxes not in the top-layer, but for box in the top layer, spec [1]
says it should use the initial containing block as the static-position
rectangle, not the rectangle where the box is placed as if it's not
absolutely positioned. Implement this behavior by resolving auto insets
for absolutely-positioned box in the top layer to 0.

[1]: https://drafts.csswg.org/css-position-3/#staticpos-rect)

* LayoutTests/imported/w3c/web-platform-tests/css/css-position/box-in-top-layer-uses-icb-as-static-position-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-position/box-in-top-layer-uses-icb-as-static-position-ref.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-position/box-in-top-layer-uses-icb-as-static-position.html: Added.
* Source/WebCore/rendering/PositionedLayoutConstraints.cpp:
(WebCore::PositionedLayoutConstraints::captureInsets):